Export works now (kinda)
Finally figured out what caused all the problems and an ugly workaround for a part of it. When creating the class the constructor doesn't set some important variables regarding context, threads and probably some more. For the context I found the workaround of just passing the context of SettingsActivity to ExportSettings and using that instead of 'this' or 'getBaseContext()'. While that works for the Shared Preferences, it does not work for the FileChooser as that gets into problems with the aforementioned threads.

But exporting works now, it saves the testFile.json to the Downloads folder. The JSON includes the settings and all devices but not the keyfiles themself.

Also added JSON.simple to dependencies and changed Copyright notice in ExportSettings.java from 2019 to 2022.