Exporting + importing is working now!
The exporting and importing of the settings and devices works now, although there is still a lot to do, especially regarding validation during the import.

The "created_at" and "modified_at" dates are no longer in exports as they can't be imported through the DeviceDbHelper's create method.

I still have no clue on how to fix the exportSettings and importSettings problems regarding them not getting context/other important stuff, thus I still can't use the file/directory selection and have to use the "passing the context" and "hardcoding the filepath" workarounds.