Validation finished, translations + fixes
Changed all hardcoded text to string values with default and German translations.

Forgot to export/import commands, that's fully added now.

Validation is also fully finished now, tested it with 2 devices and 2 commands and lot's of errors in the JSON and it seems to catch everything properly. Tested it on an emulator and on a real phone. Although more testing would definitely be nice.

Fixed ExportSettings by actually reading the DB instead of assuming device IDs always start at 1. Also it now appends a number to the filename if the file already exists and I added better error feedback.

Now all that's left is getting the filepicker to work, especially as it's needed to get file read/write permissions.